Adding folders and splitting js
