+++ app-main/public/index.html
@@ -221,7 +221,7 @@
       <script src="https://www.gstatic.com/firebasejs/7.22.0/firebase-firestore.js"></script>
       <script src="https://cdn.quilljs.com/1.3.6/quill.js"></script>
     </body>
-    <script type="module" src="./js/index.js"></script>
+    <script type="module" src="./js/index/index.js"></script>
     <!-- <script type="module" src="./js/firebase.js"></script> -->
   </body>
 </html>
