+++ app-main/public/js/firebase/firebase.js
