+++ app-main/public/js/firebase/firebase_quill_init.js
