+++ app-main/public/js/index/avatars.js
