+++ app-main/public/js/index/eventHandlers.js
@@ -1,13 +1,13 @@
 import { quill } from "../firebase/firebase_quill_init.js";
 
-exports.handleClear= function(e) {
+exports.handleClear = function (e) {
   quill.root.innerHTML = "";
   lessonInput.value = "";
   clearBtn.setAttribute("hidden", "");
   submitLessonElement.textContent = "ADD LESSON";
-}
+};
 
-exports.handleViewClick = function(lesson) {
+exports.handleViewClick = function (lesson) {
   const title = lesson.querySelector(".lesson-card-title").innerText;
   const content = lesson.querySelector(".lesson-card-content").innerHTML;
 
@@ -16,9 +16,9 @@ exports.handleViewClick = function(lesson) {
   modalLesson.removeAttribute("hidden");
   overlay.removeAttribute("hidden");
   overlay.classList.add("dark");
-}
+};
 
-exports.handleClearBtn = function() {
+exports.handleClearBtn = function () {
   let textLessonContent = quill.root.innerHTML;
   let textTitleContent = document.querySelector(".create-lesson-input");
   if (
@@ -29,15 +29,15 @@ exports.handleClearBtn = function() {
   } else {
     clearBtn.setAttribute("hidden", "");
   }
-}
+};
 
-exports.handleCloseLessonModal = function() {
+exports.handleCloseLessonModal = function () {
   modalLessonTitle.innerHTML = "";
   modalLessonContent.innerHTML = "";
 
   modalLesson.setAttribute("hidden", "");
   overlay.setAttribute("hidden", "");
   overlay.classList.remove("dark");
-}
+};
 
 module.exports = exports;
