+++ app-main/public/js/index/index.js
@@ -1,8 +1,7 @@
 // import quill init
 import { quill } from "../firebase/firebase_quill_init.js";
-import avatars from './avatars.js';
-import modal from './modals.js';
-
+import avatars from "./avatars.js";
+import modal from "./modals.js";
 
 // cached DOM elements
 const preAuthContainer = document.querySelector(".pre-auth-container");
@@ -15,7 +14,9 @@ const modalLesson = document.querySelector(".modal-lesson");
 const modalLessonClose = document.querySelector(".modal-lesson-close");
 const modalLessonTitle = document.querySelector(".modal-lesson-title");
 const modalLessonContent = document.querySelector(".modal-lesson-content");
-const createLessonContainer = document.querySelector(".create-lesson-container");
+const createLessonContainer = document.querySelector(
+  ".create-lesson-container"
+);
 const lessonInput = document.querySelector(".create-lesson-input");
 const formElement = document.querySelector("form");
 const submitLessonElement = document.querySelector("#submit");
@@ -33,7 +34,6 @@ modalLessonClose.addEventListener("click", handleCloseLessonModal);
 clearBtn.addEventListener("click", handleClear);
 overlay.addEventListener("click", handleCloseLessonModal);
 formElement.addEventListener("keyup", handleClearBtn);
-
 
 function lessonHelper({
   varName,
@@ -115,12 +115,14 @@ function handleNoLessons() {
 }
 
 function handleLessonsCount() {
-	if (JSON.parse(localStorage.getItem("user")).lessons.length) {
-		lessonCount.innerHTML = JSON.parse(localStorage.getItem("user")).lessons.length;
-	} else {
-		lessonCount.innerHTML = '';
-		return;
-	}
+  if (JSON.parse(localStorage.getItem("user")).lessons.length) {
+    lessonCount.innerHTML = JSON.parse(
+      localStorage.getItem("user")
+    ).lessons.length;
+  } else {
+    lessonCount.innerHTML = "";
+    return;
+  }
 }
 
 function handleEditClick(lesson) {
@@ -158,91 +160,91 @@ function lessonHandler(e) {
 }
 
 function renderLessons({ lessons }) {
-	handleClear();
+  handleClear();
 
-	if (lessonsContainer.childElementCount) {
-		lessonsContainer.innerHTML = "";
-	}
-	lessons.forEach(({ title, content, id }) => {
-		const lessonCard = lessonHelper({
-			varName: document.createElement("div"),
-			eventListener: { click: lessonHandler },
-			classList: ["lesson-card"],
-			attribute: [{ "data-id": id }],
-		});
-		const buttonContainer = lessonHelper({
-			varName: document.createElement("div"),
-			classList: ["lesson-card-content-buttons"],
-		});
+  if (lessonsContainer.childElementCount) {
+    lessonsContainer.innerHTML = "";
+  }
+  lessons.forEach(({ title, content, id }) => {
+    const lessonCard = lessonHelper({
+      varName: document.createElement("div"),
+      eventListener: { click: lessonHandler },
+      classList: ["lesson-card"],
+      attribute: [{ "data-id": id }],
+    });
+    const buttonContainer = lessonHelper({
+      varName: document.createElement("div"),
+      classList: ["lesson-card-content-buttons"],
+    });
 
-		const titleContainer = lessonHelper({
-			varName: document.createElement("div"),
-			classList: ["lesson-card-title-container"],
-		});
+    const titleContainer = lessonHelper({
+      varName: document.createElement("div"),
+      classList: ["lesson-card-title-container"],
+    });
 
-		const lessonTitle = lessonHelper({
-			varName: document.createElement("h2"),
-			classList: ["lesson-card-title"],
-			textContent: title,
-		});
+    const lessonTitle = lessonHelper({
+      varName: document.createElement("h2"),
+      classList: ["lesson-card-title"],
+      textContent: title,
+    });
 
-		const lessonContent = lessonHelper({
-			varName: document.createElement("div"),
-			classList: ["lesson-card-content", "ql-editor", "ql-container"],
-			innerHTML: content,
-		});
+    const lessonContent = lessonHelper({
+      varName: document.createElement("div"),
+      classList: ["lesson-card-content", "ql-editor", "ql-container"],
+      innerHTML: content,
+    });
 
-		const lessonRemoveBtn = lessonHelper({
-			varName: document.createElement("div"),
-			classList: ["button"],
-			id: "delete",
-		});
+    const lessonRemoveBtn = lessonHelper({
+      varName: document.createElement("div"),
+      classList: ["button"],
+      id: "delete",
+    });
 
-		const removeIcon = lessonHelper({
-			varName: document.createElement("img"),
-			attribute: [
-				{ alt: "remove lesson icon" },
-				{ src: "./images/cancel-white.svg" },
-			],
-			id: "delete",
-		});
+    const removeIcon = lessonHelper({
+      varName: document.createElement("img"),
+      attribute: [
+        { alt: "remove lesson icon" },
+        { src: "./images/cancel-white.svg" },
+      ],
+      id: "delete",
+    });
 
-		const editIcon = lessonHelper({
-			varName: document.createElement("img"),
-			attribute: [
-				{ alt: "edit lesson icon" },
-				{ src: "./images/edit-white.svg" },
-			],
-			id: "edit",
-		});
+    const editIcon = lessonHelper({
+      varName: document.createElement("img"),
+      attribute: [
+        { alt: "edit lesson icon" },
+        { src: "./images/edit-white.svg" },
+      ],
+      id: "edit",
+    });
 
-		const lessonEditBtn = lessonHelper({
-			varName: document.createElement("button"),
-			classList: ["button"],
-			id: "edit",
-		});
+    const lessonEditBtn = lessonHelper({
+      varName: document.createElement("button"),
+      classList: ["button"],
+      id: "edit",
+    });
 
-		const lessonViewBtn = lessonHelper({
-			varName: document.createElement("button"),
-			classList: ["button"],
-			textContent: "VIEW LESSON",
-			id: "view",
-		});
+    const lessonViewBtn = lessonHelper({
+      varName: document.createElement("button"),
+      classList: ["button"],
+      textContent: "VIEW LESSON",
+      id: "view",
+    });
 
-		titleContainer.appendChild(lessonTitle);
-		lessonRemoveBtn.appendChild(removeIcon);
-		titleContainer.appendChild(lessonRemoveBtn);
-		lessonCard.appendChild(titleContainer);
-		lessonCard.appendChild(lessonContent);
-		buttonContainer.appendChild(lessonViewBtn);
-		lessonEditBtn.appendChild(editIcon);
-		buttonContainer.appendChild(lessonEditBtn);
-		lessonCard.appendChild(buttonContainer);
-		lessonsContainer.appendChild(lessonCard);
-	});
+    titleContainer.appendChild(lessonTitle);
+    lessonRemoveBtn.appendChild(removeIcon);
+    titleContainer.appendChild(lessonRemoveBtn);
+    lessonCard.appendChild(titleContainer);
+    lessonCard.appendChild(lessonContent);
+    buttonContainer.appendChild(lessonViewBtn);
+    lessonEditBtn.appendChild(editIcon);
+    buttonContainer.appendChild(lessonEditBtn);
+    lessonCard.appendChild(buttonContainer);
+    lessonsContainer.appendChild(lessonCard);
+  });
 
-	handleNoLessons();
-	handleLessonsCount();
+  handleNoLessons();
+  handleLessonsCount();
 }
 
 function addLesson() {
